Fixed tons of warnings

+++ src/DataLoaderEasyPBR.cxx
@@ -430,7 +430,7 @@ void DataLoaderEasyPBR::read_data(){
 
 
 Frame DataLoaderEasyPBR::get_next_frame(){
-    CHECK(m_idx_img_to_read<m_frames.size()) << "m_idx_img_to_read is out of bounds. It is " << m_idx_img_to_read << " while m_frames has size " << m_frames.size();
+    CHECK(m_idx_img_to_read<(int)m_frames.size()) << "m_idx_img_to_read is out of bounds. It is " << m_idx_img_to_read << " while m_frames has size " << m_frames.size();
     Frame  frame= m_frames[m_idx_img_to_read];
 
     if(!m_do_overfit){
@@ -443,7 +443,7 @@ std::vector<easy_pbr::Frame> DataLoaderEasyPBR::get_all_frames(){
     return m_frames;
 }
 Frame DataLoaderEasyPBR::get_frame_at_idx( const int idx){
-    CHECK(idx<m_frames.size()) << "idx is out of bounds. It is " << idx << " while m_frames has size " << m_frames.size();
+    CHECK(idx<(int)m_frames.size()) << "idx is out of bounds. It is " << idx << " while m_frames has size " << m_frames.size();
 
     Frame  frame= m_frames[idx];
 
@@ -486,7 +486,7 @@ std::vector< easy_pbr::Frame > DataLoaderEasyPBR::furthest_frame_sampler( std::v
     selected_frames.push_back(frames[0]);
     is_frame_pushed[0]=true;
 
-    while(selected_frames.size()<nr_frames_to_pick){
+    while((int)selected_frames.size()<nr_frames_to_pick){
         //loop through all the input frames and also loop through the selected ones, keep an idx of the furthest one
         int idx_furthest=0;
         float maximum_distance=std::numeric_limits<float>::min();
@@ -530,11 +530,11 @@ std::vector< easy_pbr::Frame > DataLoaderEasyPBR::furthest_frame_sampler( std::v
 
 std::vector<easy_pbr::Frame>  DataLoaderEasyPBR::get_close_frames( const easy_pbr::Frame& frame, const int nr_frames, const bool discard_same_idx){
 
-    CHECK(nr_frames<m_frames.size()) << "Cannot select more close frames than the total nr of frames that we have in the loader. Required select of " << nr_frames << " out of a total of " << m_frames.size() << " available in the loader";
+    CHECK(nr_frames<(int)m_frames.size()) << "Cannot select more close frames than the total nr of frames that we have in the loader. Required select of " << nr_frames << " out of a total of " << m_frames.size() << " available in the loader";
 
     std::vector<easy_pbr::Frame> selected_close_frames;
 
-    for(size_t i=0; i<nr_frames; i++){
+    for(int i=0; i<nr_frames; i++){
 
         //select a close frame
         float closest_distance=std::numeric_limits<float>::max();
@@ -629,7 +629,7 @@ std::vector<easy_pbr::Frame>  DataLoaderEasyPBR::get_close_frames( const easy_pb
 
 bool DataLoaderEasyPBR::is_finished(){
     //check if this loader has returned all the images it has
-    if(m_idx_img_to_read<m_frames.size()){
+    if(m_idx_img_to_read<(int)m_frames.size()){
         return false; //there is still more files to read
     }
